ignore _stacklevel, make ​​parameter name​​ compatible with input and dim
