+++ test/legacy_test/test_softmax_op.py
@@ -662,5 +662,56 @@ class TestSoftmaxAPI_ZeroSize(unittest.TestCase):
             paddle.enable_static()
 
 
+class TestSoftmaxAPI_CompatibleWithTorch(TestSoftmaxAPI):
+    # torch.nn.functional.softmax(input, dim=None, _stacklevel=3, dtype=None)
+    def setUp(self):
+        self.place = get_device_place()
+        self.executed_api()
+        self.x_np_list = [
+            np.random.uniform(-1.0, 1.0, list(range(2, ndim + 2))).astype(
+                'float32'
+            )
+            for ndim in range(1, 6)
+        ]
+        self.out_ref_list = [
+            ref_softmax(x_np, axis=-1, dtype=None) for x_np in self.x_np_list
+        ]
+
+    def test_static_check(self):
+        with static_guard():
+            for func in [F.softmax, paddle.softmax, paddle.Tensor.softmax]:
+                for x_np, out_ref in zip(self.x_np_list, self.out_ref_list):
+                    with paddle.static.program_guard(paddle.static.Program()):
+                        x = paddle.static.data('X', x_np.shape, 'float32')
+                        out1 = func(input=x, dim=-1, _stacklevel=3)
+                        out2 = func(x, -1, 3)
+                        exe = paddle.static.Executor(self.place)
+                        res = exe.run(feed={'X': x_np}, fetch_list=[out1, out2])
+                        for rr in res:
+                            np.testing.assert_allclose(out_ref, rr, rtol=1e-05)
+
+    def test_dygraph_check(self):
+        paddle.disable_static(self.place)
+        for func in [F.softmax, paddle.softmax, paddle.Tensor.softmax]:
+            for x_np, out_ref in zip(self.x_np_list, self.out_ref_list):
+                x = paddle.to_tensor(x_np)
+                out1 = func(input=x, dim=-1, _stacklevel=3)
+                x = paddle.to_tensor(x_np)
+                out2 = func(x, -1, 3)
+                for r in [out1, out2]:
+                    np.testing.assert_allclose(out_ref, r.numpy(), rtol=1e-05)
+
+                # explicitly use float32 for ROCm, as MIOpen does not yet support float64
+                if core.is_compiled_with_rocm():
+                    out = func(x, dim=-1, _stacklevel=3, dtype=np.float32)
+                    out_ref = ref_softmax(x_np, axis=-1, dtype=np.float32)
+                else:
+                    out = func(x, dim=-1, _stacklevel=3, dtype=np.float64)
+                    out_ref = ref_softmax(x_np, axis=-1, dtype=np.float64)
+                np.testing.assert_allclose(out_ref, out.numpy(), rtol=1e-05)
+
+        paddle.enable_static()
+
+
 if __name__ == "__main__":
     unittest.main()
